+++ trixnity-messenger-compose-view/src/commonMain/kotlin/de/connect2x/trixnity/messenger/compose/view/i18n/I18n.kt
@@ -2031,6 +2031,16 @@ open class I18nView(
         DE - "Profilbild ändern"
     }
 
+    open fun profileAvatarDelete() = translate {
+        EN - "delete avatar"
+        DE - "Profilbild löschen"
+    }
+
+    open fun profileAvatarDeleteNoPermission() = translate {
+        EN - "You do not have permission to delete your avatar"
+        DE - "Sie haben keine Berechtigung, Ihr Profilbild zu löschen"
+    }
+
     open fun profileYourName() = translate {
         EN - "Your name"
         DE - "Ihr Name"