+++ trixnity-messenger/src/commonMain/kotlin/de/connect2x/trixnity/messenger/DI.kt
@@ -32,6 +32,7 @@ import de.connect2x.trixnity.messenger.util.Search
 import de.connect2x.trixnity.messenger.util.SearchImpl
 import de.connect2x.trixnity.messenger.util.SharedDataHandler
 import de.connect2x.trixnity.messenger.util.SharedDataHandlerImpl
+import de.connect2x.trixnity.messenger.util.audioMetadataFactoryModule
 import de.connect2x.trixnity.messenger.util.platformCloseAppModule
 import de.connect2x.trixnity.messenger.util.platformDeleteAccountDataModule
 import de.connect2x.trixnity.messenger.util.platformGetDefaultDisplayNameModule
@@ -304,6 +305,7 @@ fun createTrixnityMessengerDefaultModuleFactories(): List<ModuleFactory> =
         ::platformDeleteProfileDataModule,
         ::platformProcessImageUploadModule,
         ::platformGetImageDimensionsModule,
+        ::audioMetadataFactoryModule,
     )
 
 /*